Keep dev green by fixing Windows worktree cleanup and test-bulk budget drift
The delegate worktree cleanup path was invoking `git worktree remove` from
inside the target worktree itself. That can leave Windows holding the current
directory open, which matches the CI permission-denied failure when removing a
clean child worktree. The retained-worktree async test was also using the same
fragile removal pattern.

At the same time, `turn_coordinator.rs` had drifted past the strict
architecture line budget because the file still inlined a very large test
module. This change keeps production logic intact, moves the inline tests into a
sibling test file, and restores the file-size budget without changing runtime
behavior.

Constraint: The fix must preserve existing delegate isolation semantics while remaining a narrow baseline-ci repair
Rejected: Retry the same in-worktree removal command on Windows | keeps the root cwd-locking problem instead of removing it
Rejected: Silence the architecture budget for turn_coordinator | hides structural debt instead of reducing it
Confidence: high
Scope-risk: moderate
Reversibility: clean
Directive: Keep delegate worktree removal anchored at the owning repo root, and keep large test modules out of budgeted production files
Tested: cargo fmt --all --check; LOONGCLAW_ARCH_STRICT=true ./scripts/check_architecture_boundaries.sh
Not-tested: Full targeted cargo tests are still contending with concurrent local repo-wide rust builds; rely on fresh CI for the complete matrix
Related: #1066
