+++ crates/app/src/conversation/tests.rs
@@ -23867,22 +23867,11 @@ async fn handle_turn_with_runtime_delegate_async_worktree_isolation_retains_dirt
     assert_eq!(payloads[0]["isolation"], "worktree");
     assert_eq!(payloads[0]["workspace_retained"], true);
 
-    let git_executable = delegate_test_git_executable();
-    let cleanup_status = std::process::Command::new(git_executable)
-        .args([
-            "-C",
-            workspace_root.as_str(),
-            "worktree",
-            "remove",
-            "--force",
-            workspace_root.as_str(),
-        ])
-        .status()
-        .expect("cleanup retained worktree status");
-    assert!(
-        cleanup_status.success(),
-        "cleanup retained worktree should succeed"
-    );
+    super::workspace_isolation::cleanup_prepared_delegate_workspace_root(
+        super::ConstrainedSubagentIsolation::Worktree,
+        Some(worktree_root.as_path()),
+    )
+    .expect("cleanup retained worktree should succeed");
 }
 
 #[cfg(feature = "memory-sqlite")]
